城市选择索引 + 集成百度地图

+++ app/src/main/java/com/wjc/cateye/moive/bean/WillShowListBean.java
@@ -125,7 +125,7 @@ public class WillShowListBean {
             private String img;
             private boolean late;
             private int localPubSt;
-            private int mk;
+            private double mk;
             private String nm;
             private int pn;
             private boolean preShow;
@@ -137,7 +137,7 @@ public class WillShowListBean {
             private int recentShowDate;
             private int recentShowNum;
             private String rt;
-            private int sc;
+            private double sc;
             private String scm;
             private String showInfo;
             private int showNum;
@@ -266,11 +266,11 @@ public class WillShowListBean {
                 this.localPubSt = localPubSt;
             }
 
-            public int getMk() {
+            public double getMk() {
                 return mk;
             }
 
-            public void setMk(int mk) {
+            public void setMk(double mk) {
                 this.mk = mk;
             }
 
@@ -362,11 +362,11 @@ public class WillShowListBean {
                 this.rt = rt;
             }
 
-            public int getSc() {
+            public double getSc() {
                 return sc;
             }
 
-            public void setSc(int sc) {
+            public void setSc(double sc) {
                 this.sc = sc;
             }
